requirements.txt
pdfplumber
python-docx
docxtpl
spacy
sentence-transformers
scikit-learn
numpy
pandas
rapidfuzz
faiss-cpu
yake
rank-bm25
pyyaml
python-dotenv
requests
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-jose
passlib
bcrypt
python-multipart
python-dotenv
psycopg2-binary
aiosqlite
